+++ packages/element/src/index.ts
@@ -95,8 +95,6 @@ export * from "./textMeasurements";
 export * from "./textWrapping";
 export * from "./transform";
 export * from "./transformHandles";
-export * from "./transaction/ledger";
-export * from "./transaction/types";
 export * from "./typeChecks";
 export * from "./utils";
 export * from "./zindex";
